+++ CHANGELOG.md
@@ -8,9 +8,11 @@ A more detailed list of changes is available in the corresponding milestones for
   - also improved the readability of **com.google.fonts/check/points_out_of_bounds**
 
 ### New Checks
+  - **[com.google.fonts/check/description/git_url]:** Make sure all font families have an upstream git repo URL declared in the DESCRIPTION.en-us.html file
   - **[com.google.fonts/check/varfont_instance_coordinates]:** Check variable font instances have correct axis coordinates
   - **[com.google.fonts/check/varfont_instance_names]:** Check variable font instances have correct names
   - **[com.google.fonts/check/wdth_valid_range]:** Check variable font wdth axis has correct range
+
 
 ## 0.7.5 (2019-May-24)
 ### Note-worthy code changes